fix: use float for speed in animation phase calculations
getSpeed() returns float (m_speed / 100.f), but was being stored as
const int, truncating values < 1.0 to 0. This triggered the speed <= 0
guard and killed animations for any attached effect with speed < 100
(e.g. wings/auras configured with speed = 0.5).

Also fixes thingTye typo to thingType in attachedeffect.cpp.

+++ src/client/paperdoll.cpp
@@ -118,8 +118,8 @@ int Paperdoll::getCurrentAnimationPhase()
 
     if (m_thingType->isCreature() && m_thingType->isAnimateAlways()) {
         const int animationPhases = m_thingType->getAnimationPhases();
-        const int speed = getSpeed();
-        if (animationPhases <= 0 || speed <= 0) return 0;
+        const float speed = getSpeed();
+        if (animationPhases <= 0 || speed <= 0.f) return 0;
 
         const int ticksPerFrame = std::max<int>(1, static_cast<int>(std::round((1000.0 / animationPhases) / speed)));
         const long long animationPeriod = static_cast<long long>(ticksPerFrame) * animationPhases;
